# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MangaItem {
  id: string;
  title: string;
  url: string | null;
  image: string | null;
}

export interface MangaListResponse {
  mangas: MangaItem[];
  total: number;
  page: number;
  limit: number;
}

export interface ScrapeStatus {
  lastScraped: string | null;
  total: number;
  isRunning: boolean;
  nextScheduled: string | null;
}

export interface ScrapeResult {
  success: boolean;
  count: number;
  timestamp: string;
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ListMangaParams = {
page?: number;
limit?: number;
};

export type SearchMangaParams = {
q: string;
};