# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: manga
    description: Manga operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /manga:
    get:
      operationId: listManga
      tags: [manga]
      summary: List all manga
      description: Returns all scraped manga entries with pagination
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of manga
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaListResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /manga/status:
    get:
      operationId: getMangaStatus
      tags: [manga]
      summary: Get scrape status
      description: Returns the last scrape timestamp and total count
      responses:
        "200":
          description: Scrape status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeStatus"

  /manga/search:
    get:
      operationId: searchManga
      tags: [manga]
      summary: Search manga by title
      description: Returns manga matching the search query
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaListResponse"
        "400":
          description: Missing query parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /manga/{id}:
    get:
      operationId: getManga
      tags: [manga]
      summary: Get a single manga by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Single manga entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /scrape:
    post:
      operationId: triggerScrape
      tags: [manga]
      summary: Manually trigger a scrape
      description: Starts the scraper and returns updated results
      responses:
        "200":
          description: Scrape result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeResult"
        "409":
          description: Scrape already in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Scrape failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    MangaItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        url:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
      required:
        - id
        - title
        - url
        - image

    MangaListResponse:
      type: object
      properties:
        mangas:
          type: array
          items:
            $ref: "#/components/schemas/MangaItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - mangas
        - total
        - page
        - limit

    ScrapeStatus:
      type: object
      properties:
        lastScraped:
          type: string
          nullable: true
        total:
          type: integer
        isRunning:
          type: boolean
        nextScheduled:
          type: string
          nullable: true
      required:
        - lastScraped
        - total
        - isRunning
        - nextScheduled

    ScrapeResult:
      type: object
      properties:
        success:
          type: boolean
        count:
          type: integer
        timestamp:
          type: string
        message:
          type: string
      required:
        - success
        - count
        - timestamp
        - message

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message