# lib/api-zod/src/generated/types/mangaItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MangaItem {
  id: string;
  title: string;
  url: string | null;
  image: string | null;
}