# lib/api-zod/src/generated/types/mangaListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MangaItem } from './mangaItem';

export interface MangaListResponse {
  mangas: MangaItem[];
  total: number;
  page: number;
  limit: number;
}