# lib/api-zod/src/generated/types/scrapeResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScrapeResult {
  success: boolean;
  count: number;
  timestamp: string;
  message: string;
}