# lib/api-zod/src/generated/types/scrapeStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScrapeStatus {
  lastScraped: string | null;
  total: number;
  isRunning: boolean;
  nextScheduled: string | null;
}